client: Add a tokio_layer module connecting solicit to Tokio
This module mainly exposes an implementation of Tokio's Transport trait,
which allows it to use it as a Sink for request frames (request headers,
request body chunks), as well as as a Stream of response frames
(response headers, response body chunks).

Currently, the full response to the request will always appear
"bodiless" to Tokio, as the response "headers" struct will carry both
the HTTP response headers, as well as the full body.

In the future, when we start yielding Body frames from the Transport
stream, Tokio will automatically hook up each resolved response future
with a `Stream` impl that will be receiving these chunks, as long as
there are still any.