Implement response body streaming
In order to allow response bodies to be properly streamed, we need to
yield body *chunks* to Tokio. This change implements this by adding an
implementation of the `solicit` `Stream` trait and using this instead of
the `DefaultStream`.

`solicit` `Stream`s are an abstraction over the HTTP/2 stream concept.
They allow us to hook into events that arise on individual h2 streams.
At this layer, we don't need to care about general HTTP/2 session events,
but only how to hook up streams to requests -- which is exactly what we
need to connect to Tokio.

The `Stream` implementation will communicate three kinds of messages back
to `H2ClientTokioTransport`. Firstly, it communicates when a header
becomes available; secondly, for each incoming data chunk, it posts a
response body chunk. `H2ClientTokioTransport` converts these messages to
appropriate `Frame`s that it can yield to Tokio, which takes care of
adapting them into a `futures` `Stream`, representing the body. Lastly,
it notifies the transport when the stream transitions into a
remote-closed state, meaning that no more data chunks will be coming.

The `H2Client` is also adapted to now yield futures with both the
response headers, as well as a `futures` `Stream` of the body.

The example is updated accordingly.