client: Refactor returned Future to be a wrapper struct
Currently, the Future that the client returns is a simple type alias
for a boxed future. This commit makes it a full struct, but still keeps
it a simple wrapper around a boxed future.

This way, the representation of the future can evolve without leaking
details of how it's internally obtained/implemented.