client: Add a convenience method to accumulate the response body
The future representing the response now represents the body as a
stream of chunks. However, often it is useful to simply wait for the
entire body to come in before doing anything, which is now facilitated
by the convenience method on the returned `Future`.

It simply accumulates all the body chunks into a `Vec` and returns a
boxed future with both the original headers and the accumulated Vec.