Properly handle errors, removing `expect`s
The various error handling that was ignored by simply `expect`ing on
the returned `Result`s is now fixed, such that the errors are
propagated out.

Currently, Tokio seems to require that the Transport (and its
associated Sink + Stream traits) use `io::Error`. Therefore, at this
boundary, we convert to an `io::Error` instance that will wrap a
potential h2 protocol error.

Internally, the Transport uses a new `Http2Error`, which can be
implicitly created from the `solicit::http::HttpError`, as well as an
`io::Error`.