Refactor the proto to use a connector to negotiate the connection
Negotiating the use of http/2 on the application layer can be done in a
few different ways. The concept of the Connector can now be used to
provide a custom implementation for this step.

This commit includes a `CleartextConnector` which simply assumes that
the server will accept the use of an http/2 connection over the
cleartext TCP socket.